Separated generic KeyframeAnimation from PartAnimation code.
